- Subquery Jpql xử lí left join expect 1 row join (group by condition)

+++ BE/src/main/java/com/minh/entity/id/CompositeId.java
@@ -10,9 +10,9 @@ public class CompositeId implements Serializable {
 
     private Long commentId;
 
-    private Long viewerId;
+    private Integer viewerId;
 
-    public CompositeId(Long commentId, Long viewerId) {
+    public CompositeId(Long commentId, Integer viewerId) {
         this.commentId = commentId;
         this.viewerId = viewerId;
     }
